+++ apps/backoffice-v2/src/pages/TransactionMonitoringAlertsAnalysis/components/AlertAnalysisSheet/columns.tsx
@@ -55,8 +55,8 @@ export const columns = [
         <div className={`w-[11.8ch]`}>
           <TextWithNAFallback
             className={buttonVariants({
-              variant: 'link',
-              className: '!block cursor-pointer !p-0 text-sm !text-blue-500',
+              variant: 'ghost',
+              className: '!block !p-0 text-sm',
             })}
             style={styles}
             ref={ref}
@@ -89,9 +89,7 @@ export const columns = [
       const counterpartyOriginatorName = info.getValue();
 
       return (
-        <TextWithNAFallback className="text-sm font-semibold">
-          {counterpartyOriginatorName}
-        </TextWithNAFallback>
+        <TextWithNAFallback className="text-sm">{counterpartyOriginatorName}</TextWithNAFallback>
       );
     },
     header: 'Originator Name',
@@ -101,7 +99,7 @@ export const columns = [
       const counterpartyOriginatorCorrelationId = info.getValue();
 
       return (
-        <TextWithNAFallback className="text-sm font-semibold">
+        <TextWithNAFallback className="text-sm">
           {counterpartyOriginatorCorrelationId}
         </TextWithNAFallback>
       );
@@ -113,9 +111,7 @@ export const columns = [
       const counterpartyBeneficiaryName = info.getValue();
 
       return (
-        <TextWithNAFallback className="text-sm font-semibold">
-          {counterpartyBeneficiaryName}
-        </TextWithNAFallback>
+        <TextWithNAFallback className="text-sm">{counterpartyBeneficiaryName}</TextWithNAFallback>
       );
     },
     header: 'Beneficiary Name',
@@ -125,7 +121,7 @@ export const columns = [
       const counterpartyBeneficiaryCorrelationId = info.getValue();
 
       return (
-        <TextWithNAFallback className="text-sm font-semibold">
+        <TextWithNAFallback className="text-sm">
           {counterpartyBeneficiaryCorrelationId}
         </TextWithNAFallback>
       );
@@ -137,9 +133,7 @@ export const columns = [
       const paymentMethod = info.getValue() ?? '';
 
       return (
-        <TextWithNAFallback className="text-sm font-semibold">
-          {titleCase(paymentMethod)}
-        </TextWithNAFallback>
+        <TextWithNAFallback className="text-sm">{titleCase(paymentMethod)}</TextWithNAFallback>
       );
     },
     header: 'Payment Method',