+++ services/workflows-service/src/data-analytics/types.ts
@@ -1,5 +1,7 @@
+import { TProjectId } from '@/types';
 import { TransactionDirection, PaymentMethod, TransactionRecordType } from '@prisma/client';
 import { AggregateType, TIME_UNITS } from './consts';
+
 export type InlineRule = {
   id: string;
   subjects: string[] | readonly string[];
@@ -17,8 +19,28 @@ export type InlineRule = {
       options: Omit<TCustomersTransactionTypeOptions, 'projectId'>;
     }
   | {
-      fnName: 'evaluateTransactionLimitHistoricAverageInbound';
+      fnName: 'evaluateTransactionAvg';
       options: Omit<TransactionLimitHistoricAverageOptions, 'projectId'>;
+    }
+  | {
+      fnName: 'evaluateTransactionAvg';
+      options: Omit<TPeerGroupTransactionAverageOptions, 'projectId'>;
+    }
+  | {
+      fnName: 'evaluateDormantAccount';
+      options: Omit<TDormantAccountOptions, 'projectId'>;
+    }
+  | {
+      fnName: 'checkMerchantOngoingAlert';
+      options: CheckRiskScoreOptions;
+    }
+  | {
+      fnName: 'evaluateHighVelocityHistoricAverage';
+      options: Omit<HighVelocityHistoricAverageOptions, 'projectId'>;
+    }
+  | {
+      fnName: 'evaluateMultipleMerchantsOneCounterparty';
+      options: Omit<TMultipleMerchantsOneCounterparty, 'projectId'>;
     }
 );
 
@@ -32,7 +54,7 @@ export type TExcludedCounterparty = {
 export type TimeUnit = (typeof TIME_UNITS)[keyof typeof TIME_UNITS];
 
 export type TransactionsAgainstDynamicRulesType = {
-  projectId: string;
+  projectId: TProjectId;
   havingAggregate?: TAggregations;
   amountBetween?: { min: number; max: number };
   timeAmount?: number;
@@ -48,7 +70,7 @@ export type TransactionsAgainstDynamicRulesType = {
 };
 
 export type HighTransactionTypePercentage = {
-  projectId: string;
+  projectId: TProjectId;
   transactionType: TransactionRecordType;
   subjectColumn: 'counterpartyOriginatorId' | 'counterpartyBeneficiaryId';
   minimumCount: number;
@@ -58,7 +80,7 @@ export type HighTransactionTypePercentage = {
 };
 
 export type TCustomersTransactionTypeOptions = {
-  projectId: string;
+  projectId: TProjectId;
   transactionType?: TransactionRecordType[] | readonly TransactionRecordType[];
   threshold?: number;
   paymentMethods?: PaymentMethod[] | readonly PaymentMethod[];
@@ -69,7 +91,7 @@ export type TCustomersTransactionTypeOptions = {
 };
 
 export type TransactionLimitHistoricAverageOptions = {
-  projectId: string;
+  projectId: TProjectId;
   transactionDirection: TransactionDirection;
   paymentMethod: {
     value: PaymentMethod;
@@ -78,4 +100,48 @@ export type TransactionLimitHistoricAverageOptions = {
   minimumCount: number;
   minimumTransactionAmount: number;
   transactionFactor: number;
+};
+
+export type CheckRiskScoreOptions = {
+  increaseRiskScorePercentage?: number;
+  increaseRiskScore?: number;
+  maxRiskScoreThreshold?: number;
+};
+
+export type TPeerGroupTransactionAverageOptions = TransactionLimitHistoricAverageOptions & {
+  customerType?: string;
+  timeUnit?: TimeUnit;
+  timeAmount?: number;
+};
+
+export type TDormantAccountOptions = {
+  projectId: TProjectId;
+  timeAmount: number;
+  timeUnit: TimeUnit;
+};
+
+export type HighVelocityHistoricAverageOptions = {
+  projectId: TProjectId;
+  transactionDirection: TransactionDirection;
+  transactionFactor: number;
+  minimumCount: number;
+  paymentMethod: {
+    value: PaymentMethod;
+    operator: '=' | '!=';
+  };
+  activeUserPeriod: {
+    timeAmount: number;
+  };
+  lastDaysPeriod: {
+    timeAmount: number;
+  };
+  timeUnit: TimeUnit;
+};
+
+export type TMultipleMerchantsOneCounterparty = {
+  projectId: TProjectId;
+  excludedCounterparty?: TExcludedCounterparty;
+  minimumCount: number;
+  timeAmount: number;
+  timeUnit: TimeUnit;
 };