+++ services/workflows-service/src/webhooks/webhooks.controller.ts
@@ -8,10 +8,7 @@ import { WebhooksService } from '@/webhooks/webhooks.service';
 import { VerifyUnifiedApiSignatureDecorator } from '@/common/decorators/verify-unified-api-signature.decorator';
 import { BadRequestException } from '@nestjs/common';
 import { isObject } from '@ballerine/common';
-
-const Webhook = {
-  AML_INDIVIDUAL_MONITORING_UPDATE: 'aml.individuals.monitoring.update',
-} as const;
+import { AppLoggerService } from '@/common/app-logger/app-logger.service';
 
 const EntityType = {
   BUSINESS: 'business',
@@ -22,7 +19,10 @@ const EntityType = {
 @swagger.ApiTags('Webhooks')
 @common.Controller('webhooks')
 export class WebhooksController {
-  constructor(private readonly webhooksService: WebhooksService) {}
+  constructor(
+    private readonly webhooksService: WebhooksService,
+    private readonly logger: AppLoggerService,
+  ) {}
 
   @common.Post('/:entityType/aml')
   @swagger.ApiOkResponse()
@@ -32,7 +32,7 @@ export class WebhooksController {
   @VerifyUnifiedApiSignatureDecorator()
   async amlHook(
     @common.Param() { entityType }: AmlWebhookInput,
-    @common.Body() { eventName, data }: IndividualAmlWebhookInput,
+    @common.Body() { data }: IndividualAmlWebhookInput,
   ) {
     if (!(isObject(data) && 'endUserId' in data && data.endUserId)) {
       throw new BadRequestException('Missing endUserId');
@@ -40,12 +40,14 @@ export class WebhooksController {
 
     try {
       if (entityType === EntityType.INDIVIDUAL) {
-        if (eventName === Webhook.AML_INDIVIDUAL_MONITORING_UPDATE) {
-          await this.webhooksService.handleIndividualAmlHit({ endUserId: data.endUserId, data });
-        }
+        await this.webhooksService.handleIndividualAmlHit({ endUserId: data.endUserId, data });
+      } else {
+        this.logger.error(`Unknown entity type: ${entityType}`);
+
+        throw new BadRequestException('Unknown entity type');
       }
     } catch (error) {
-      console.error(error);
+      this.logger.error('amlHook::', { entityType, data, error });
 
       throw error;
     }