+++ services/workflows-service/src/webhooks/webhooks.service.ts
@@ -4,6 +4,8 @@ import { EndUserRepository } from '@/end-user/end-user.repository';
 import { CustomerService } from '@/customer/customer.service';
 import { WorkflowDefinitionService } from '@/workflow-defintion/workflow-definition.service';
 import { WorkflowService } from '@/workflow/workflow.service';
+import { AppLoggerService } from '@/common/app-logger/app-logger.service';
+import { EndUserService } from '@/end-user/end-user.service';
 
 @Injectable()
 export class WebhooksService {
@@ -12,6 +14,8 @@ export class WebhooksService {
     private readonly workflowService: WorkflowService,
     private readonly endUserRepository: EndUserRepository,
     private readonly workflowDefinitionService: WorkflowDefinitionService,
+    private readonly logger: AppLoggerService,
+    private readonly endUserService: EndUserService,
   ) {}
 
   async handleIndividualAmlHit({
@@ -21,6 +25,8 @@ export class WebhooksService {
     endUserId: string;
     data: IndividualAmlWebhookInput['data'];
   }) {
+    this.logger.log('Started handling individual AML hit', { endUserId });
+
     const { projectId, ...rest } = await this.endUserRepository.findByIdUnscoped(endUserId, {
       select: {
         approvalState: true,
@@ -57,6 +63,8 @@ export class WebhooksService {
     });
 
     if (!config?.ongoingWorkflowDefinitionId) {
+      this.logger.error('No ongoing workflow definition found for project', { projectId });
+
       return;
     }
 
@@ -65,7 +73,26 @@ export class WebhooksService {
       [projectId],
     );
 
-    await this.workflowService.createOrUpdateWorkflowRuntime({
+    const hits = (data as { hits: Array<Record<PropertyKey, unknown>> })?.hits ?? [];
+
+    const amlHits = hits.map(hit => ({
+      ...hit,
+      vendor: 'veriff',
+    }));
+
+    await this.endUserService.updateById(endUserId, {
+      data: {
+        amlHits,
+      },
+    });
+
+    if (hits.length === 0) {
+      this.logger.log('No AML hits found', { endUserId });
+
+      return;
+    }
+
+    const workflow = await this.workflowService.createOrUpdateWorkflowRuntime({
       workflowDefinitionId,
       context: {
         aml: data,
@@ -82,5 +109,7 @@ export class WebhooksService {
       projectIds: [projectId],
       currentProjectId: projectId,
     });
+
+    this.logger.log(`Created workflow for AML hits`, { workflow });
   }
 }