+++ apps/backoffice-v2/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/backoffice-v2
 
+## 0.7.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/blocks@0.2.3
+
+## 0.7.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.5
+  - @ballerine/workflow-browser-sdk@0.6.8
+  - @ballerine/workflow-node-sdk@0.6.8
+
 ## 0.7.6
 
 ### Patch Changes